trading/requirements.txt
requests~=2.32.0
beautifulsoup4~=4.13.0
alpacapy~=0.39.0
keyring~=25.6.0
pandas~=2.2.0